+++ lib/tree_haver/backends/parslet.rb
@@ -558,4 +558,3 @@ module TreeHaver
     end
   end
 end
-